renames and exports cleanup
