+++ packages/dds/tree/src/index.ts
@@ -3,16 +3,17 @@
  * Licensed under the MIT License.
  */
 
-export {
-    FieldKey,
-    TreeType,
-} from "./tree";
+export { FieldKey, TreeType } from "./tree";
+
+export { ITreeCursor, TreeNavigationResult } from "./forest";
 
 export {
-    ITreeCursor,
-    TreeNavigationResult,
-} from "./forest";
-
-export { Brand, BrandedType, Opaque, extractFromOpaque } from "./brand";
-
-export { MakeNominal } from "./typeCheck";
+    Brand,
+    BrandedType,
+    Opaque,
+    extractFromOpaque,
+    MakeNominal,
+    Invariant,
+    Contravariant,
+    Covariant,
+} from "./util";
